put everything back in the original file, made schemboundmapper
