+++ docs/samples/Microsoft.ML.Samples/Dynamic/FeatureContributionCalculationTransform.cs
@@ -67,7 +67,7 @@ namespace Microsoft.ML.Samples.Dynamic
             //var featureContributionCalculator = new FeatureContributionCalculatingEstimator(mlContext, args, model.FeatureColumn, model.Model);
             //var outputData = featureContributionCalculator.Fit(transformedData).Transform(transformedData);
 
-            var outputData1 = new FeatureContributionCalculationTransform(mlContext, new FeatureContributionCalculationTransform.Arguments() { Top = 11, Normalize = false }, model.FeatureColumn, model.Model).Transform(transformedData);
+            //var outputData1 = new FeatureContributionCalculationTransform(mlContext, new FeatureContributionCalculationTransform.Arguments() { Top = 11, Normalize = false }, model.FeatureColumn, model.Model).Transform(transformedData);
 
             var featureContributionCalculator = new FeatureContributionCalculatingTransformer(mlContext, model.Model, model.FeatureColumn, args);
             var outputData = featureContributionCalculator.Transform(transformedData);
