+++ src/Microsoft.ML.Data/Scorers/FeatureContributionCalculationTransform.cs
@@ -17,14 +17,20 @@ using Microsoft.ML.Runtime.Internal.Utilities;
 using Microsoft.ML.Runtime.Model;
 using Microsoft.ML.Runtime.Numeric;
 
-[assembly: LoadableClass(typeof(IDataScorerTransform), typeof(FeatureContributionCalculationTransform), typeof(FeatureContributionCalculationTransform.Arguments),
+[assembly: LoadableClass(typeof(IDataScorerTransform), typeof(FeatureContributionCalculatingTransformer.BindableMapper), typeof(FeatureContributionCalculatingTransformer.Arguments),
     typeof(SignatureDataScorer), "Feature Contribution Transform", "fct", "FeatureContributionCalculationTransform", MetadataUtils.Const.ScoreColumnKind.FeatureContribution)]
 
-[assembly: LoadableClass(typeof(ISchemaBindableMapper), typeof(FeatureContributionCalculationTransform), typeof(FeatureContributionCalculationTransform.Arguments),
+[assembly: LoadableClass(typeof(ISchemaBindableMapper), typeof(FeatureContributionCalculatingTransformer.BindableMapper), typeof(FeatureContributionCalculatingTransformer.Arguments),
     typeof(SignatureBindableMapper), "Feature Contribution Mapper", "fct", MetadataUtils.Const.ScoreColumnKind.FeatureContribution)]
 
-[assembly: LoadableClass(typeof(ISchemaBindableMapper), typeof(FeatureContributionCalculationTransform), null, typeof(SignatureLoadModel),
-    "Feature Contribution Mapper", FeatureContributionCalculationTransform.MapperLoaderSignature)]
+[assembly: LoadableClass(typeof(ISchemaBindableMapper), typeof(FeatureContributionCalculatingTransformer.BindableMapper), null, typeof(SignatureLoadModel),
+    "Feature Contribution Mapper", FeatureContributionCalculatingTransformer.MapperLoaderSignature)]
+
+[assembly: LoadableClass(FeatureContributionCalculatingTransformer.Summary, typeof(FeatureContributionCalculatingTransformer), null, typeof(SignatureLoadModel),
+    FeatureContributionCalculatingTransformer.FriendlyName, FeatureContributionCalculatingTransformer.LoaderSignature)]
+
+[assembly: LoadableClass(typeof(IRowMapper), typeof(FeatureContributionCalculatingTransformer), null, typeof(SignatureLoadRowMapper),
+   FeatureContributionCalculatingTransformer.FriendlyName, FeatureContributionCalculatingTransformer.LoaderSignature)]
 
 namespace Microsoft.ML.Runtime.Data
 {
@@ -43,7 +49,7 @@ namespace Microsoft.ML.Runtime.Data
     /// ]]>
     /// </format>
     /// </example>
-    public sealed class FeatureContributionCalculationTransform : ITransformer, ICanSaveModel
+    public sealed class FeatureContributionCalculatingTransformer : RowToRowTransformerBase
     {
         public sealed class Arguments : ScorerArgumentsBase
         {
@@ -68,13 +74,16 @@ namespace Microsoft.ML.Runtime.Data
         internal const string LoaderSignature = "FeatureContribution";
 
         internal const string MapperLoaderSignature = "WTFBindable";
+
         private const int MaxTopBottom = 1000;
 
-        public bool IsRowToRowMapper => true;
-
-        private readonly IHost _host;
-        private readonly BindableMapper _mapper;
         private readonly string _features;
+        private readonly int _topContributionsCount;
+        private readonly int _bottomContributionsCount;
+        private readonly bool _normalize;
+        private readonly bool _stringify;
+        private readonly IFeatureContributionMapper _predictor;
+        private readonly BindableMapper _mapper;
 
         private static VersionInfo GetVersionInfo()
         {
@@ -84,46 +93,60 @@ namespace Microsoft.ML.Runtime.Data
                 verReadableCur: 0x00010001,
                 verWeCanReadBack: 0x00010001,
                 loaderSignature: LoaderSignature,
-                loaderAssemblyName: typeof(FeatureContributionCalculationTransform).Assembly.FullName);
+                loaderAssemblyName: typeof(FeatureContributionCalculatingTransformer).Assembly.FullName);
         }
 
-        // TODO: Only thing that we need to do here is store the args, and predictor, change order of params + mew arguments object
-        public FeatureContributionCalculationTransform(IHostEnvironment env, Arguments args, string features, IPredictor predictor)
+        // TODO documentation
+        public FeatureContributionCalculatingTransformer(IHostEnvironment env, IPredictor predictor, string featuresColumn, Arguments args)
+            : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(FeatureContributionCalculatingTransformer)))
         {
-            // TODO checks that they are not empty --> features
-            // TODO change the arguments to something else --> advanced arguemnts + accessible settings
-            Contracts.CheckValue(env, nameof(env));
-            _host = env.Register(nameof(FeatureContributionCalculationTransform));
-            env.CheckValue(args, nameof(args));
-            env.CheckValue(predictor, nameof(predictor));
-            if (args.Top <= 0 || args.Top > MaxTopBottom)
-                throw env.Except($"Number of top contribution must be in range (0,{MaxTopBottom}]");
-            if (args.Bottom <= 0 || args.Bottom > MaxTopBottom)
-                throw env.Except($"Number of bottom contribution must be in range (0,{MaxTopBottom}]");
+            Host.CheckValue(args, nameof(args));
+            Host.CheckValue(predictor, nameof(predictor));
 
-            _mapper = CreateBindableMapper(env, args, predictor);
-            _features = features;
+            var pred = predictor as IFeatureContributionMapper;
+            Host.CheckParam(pred != null, nameof(predictor), "Predictor doesn't support getting feature contributions");
+
+            // TODO check that the featues column is not empty.
+            _mapper = new BindableMapper(Host, pred, args.Top, args.Bottom, args.Normalize, args.Stringify);
+            _features = featuresColumn;
+            _predictor = pred;
+            _stringify = args.Stringify;
+            _topContributionsCount = args.Top;
+            _bottomContributionsCount = args.Bottom;
+            _normalize = args.Normalize;
         }
 
-        private FeatureContributionCalculationTransform(IHostEnvironment env, ModelLoadContext ctx)
+        // Factory method for SignatureLoadModel
+        private FeatureContributionCalculatingTransformer(IHostEnvironment env, ModelLoadContext ctx)
+            : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(FeatureContributionCalculatingTransformer)))
         {
-            Contracts.CheckValue(env, nameof(env));
-            _host = env.Register(nameof(FeatureContributionCalculationTransform));
+            Host.CheckValue(ctx, nameof(ctx));
             ctx.CheckAtModel(GetVersionInfo());
 
             // *** Binary format ***
             // string features
             // BindableMapper mapper
 
+            // TODO use ctx.LoadModel with BindableMapper instead of this.
             _features = ctx.LoadNonEmptyString();
-            _mapper = new BindableMapper(env, ctx);
+            ctx.LoadModel<IFeatureContributionMapper, SignatureLoadModel>(env, out _predictor, ModelFileUtils.DirPredictor);
+            _topContributionsCount = ctx.Reader.ReadInt32();
+            Contracts.CheckDecode(0 < _topContributionsCount && _topContributionsCount <= MaxTopBottom);
+            _bottomContributionsCount = ctx.Reader.ReadInt32();
+            Contracts.CheckDecode(0 < _bottomContributionsCount && _bottomContributionsCount <= MaxTopBottom);
+            _normalize = ctx.Reader.ReadBoolByte();
+            _stringify = ctx.Reader.ReadBoolByte();
+
+            _mapper = new BindableMapper(env, _predictor, _topContributionsCount, _bottomContributionsCount, _normalize, _stringify);
         }
 
-        // TODO: we might need to save features as well..
-        public void Save(ModelSaveContext ctx)
+        // Factory method for SignatureLoadRowMapper.
+        internal static IRowMapper Create(IHostEnvironment env, ModelLoadContext ctx, ISchema inputSchema)
+            => new FeatureContributionCalculatingTransformer(env, ctx).MakeRowMapper(Schema.Create(inputSchema));
+
+        public override void Save(ModelSaveContext ctx)
         {
-            _host.CheckValue(ctx, nameof(ctx));
-            ctx.CheckAtModel();
+            Host.CheckValue(ctx, nameof(ctx));
             ctx.SetVersionInfo(GetVersionInfo());
 
             // *** Binary format ***
@@ -131,102 +154,114 @@ namespace Microsoft.ML.Runtime.Data
             // BindableMapper mapper
 
             ctx.SaveNonEmptyString(_features);
-            _mapper.Save(ctx);
+            // TODO use ctx.SaveModel with BindableMapper instead of this.
+            ctx.SaveModel(_predictor, ModelFileUtils.DirPredictor);
+            Contracts.Assert(0 < _topContributionsCount && _topContributionsCount <= MaxTopBottom);
+            ctx.Writer.Write(_topContributionsCount);
+            Contracts.Assert(0 < _bottomContributionsCount && _bottomContributionsCount <= MaxTopBottom);
+            ctx.Writer.Write(_bottomContributionsCount);
+            ctx.Writer.WriteBoolByte(_normalize);
+            ctx.Writer.WriteBoolByte(_stringify);
         }
 
-        public Schema GetOutputSchema(Schema inputSchema)
+        protected override IRowMapper MakeRowMapper(Schema schema)
+            => new Mapper(this, schema);
+
+        private class Mapper : MapperBase
         {
-            var input = new EmptyDataView(_host, inputSchema);
-            var roles = new List<KeyValuePair<RoleMappedSchema.ColumnRole, string>>();
-            roles.Add(new KeyValuePair<RoleMappedSchema.ColumnRole, string>(RoleMappedSchema.ColumnRole.Feature, _features));
-            var schema = new RoleMappedSchema(input.Schema, roles);
+            private readonly FeatureContributionCalculatingTransformer _parent;
+            private readonly BindableMapper _bindableMapper;
+            private readonly RoleMappedSchema _roleMappedSchema;
+            private readonly ISchemaBoundRowMapper _genericRowMapper;
+            private readonly Schema _outputGenericSchema;
+            private readonly VBuffer<ReadOnlyMemory<char>> _slotNames;
 
-            var boundMapper = _mapper.Bind(_host, schema);
-            var data = Create(_host, null, input, boundMapper, null);
-            return data.Schema;
+            public Mapper(FeatureContributionCalculatingTransformer parent, Schema schema)
+                : base(parent.Host, schema)
+            {
+                // TODO some checks? get soem of the columns, initialize some stuff
+                _parent = parent;
+                _bindableMapper = _parent._mapper;
+
+                var roles = new List<KeyValuePair<RoleMappedSchema.ColumnRole, string>>();
+                roles.Add(new KeyValuePair<RoleMappedSchema.ColumnRole, string>(RoleMappedSchema.ColumnRole.Feature, _parent._features));
+                _roleMappedSchema = new RoleMappedSchema(InputSchema, roles);
+
+                var genericMapper = _bindableMapper.GenericMapper.Bind(Host, _roleMappedSchema);
+                _genericRowMapper = genericMapper as ISchemaBoundRowMapper;
+                _outputGenericSchema = _genericRowMapper.OutputSchema;
+
+                if (InputSchema.HasSlotNames(_roleMappedSchema.Feature.Index, _roleMappedSchema.Feature.Type.VectorSize))
+                    InputSchema.GetMetadata(MetadataUtils.Kinds.SlotNames, _roleMappedSchema.Feature.Index,
+                        ref _slotNames);
+                else
+                    _slotNames = VBufferUtils.CreateEmpty<ReadOnlyMemory<char>>(_roleMappedSchema.Feature.Type.VectorSize);
+            }
+
+            /// <summary>
+            /// Returns the input columns needed for the requested output columns.
+            /// </summary>
+            public override Func<int, bool> GetDependencies(Func<int, bool> activeOutput)
+            {
+                var active = new bool[InputSchema.ColumnCount];
+                InputSchema.TryGetColumnIndex(_parent._features, out int featureCol);
+                active[featureCol] = true;
+                return col => active[col];
+            }
+
+            public override void Save(ModelSaveContext ctx)
+                => _parent.Save(ctx);
+
+            // The FeatureContributionCalculatingTransformer produces two columns: Score and FeatureContribution.
+            // If the argument stringify is true, the type of the FeatureContribution column is string, otherwise it is a vector of float.
+            protected override Schema.DetachedColumn[] GetOutputColumnsCore()
+            {
+                var result = new List<Schema.DetachedColumn>();
+
+                // Add Score Column.
+                foreach (var pair in _outputGenericSchema.GetColumns())
+                    result.Add(new Schema.DetachedColumn(pair.column));
+
+                // Add FeatureContributions column.
+                var builder = new MetadataBuilder();
+                builder.Add(InputSchema[_roleMappedSchema.Feature.Index].Metadata, x => x == MetadataUtils.Kinds.SlotNames);
+                if (_bindableMapper.Stringify)
+                    result.Add(new Schema.DetachedColumn(DefaultColumnNames.FeatureContributions, TextType.Instance, builder.GetMetadata()));
+                else
+                    result.Add(new Schema.DetachedColumn(DefaultColumnNames.FeatureContributions, new VectorType(NumberType.R4, _roleMappedSchema.Feature.Type.AsVector), builder.GetMetadata()));
+
+                return result.ToArray();
+            }
+
+            protected Delegate GetScoreGetter(Row input, int iinfo, Func<int, bool> activeOutput, out Action disposer)
+            {
+                var genericRow = _genericRowMapper.GetRow(input, col => activeOutput(col), out disposer);
+                return RowCursorUtils.GetGetterAsDelegate(genericRow, iinfo);
+            }
+
+            protected Delegate GetFeatureContributionGetter(Row input, int iinfo, Func<int, bool> activeOutput, out Action disposer)
+            {
+                disposer = null;
+
+                if (activeOutput(iinfo))
+                {
+                    return _bindableMapper.Stringify
+                        ? _bindableMapper.GetTextContributionGetter(input, _roleMappedSchema.Feature.Index, _slotNames)
+                        : _bindableMapper.GetContributionGetter(input, _roleMappedSchema.Feature.Index);
+                }
+                return null;
+            }
+
+            protected override Delegate MakeGetter(Row input, int iinfo, Func<int, bool> activeOutput, out Action disposer)
+            {
+                if (iinfo < _outputGenericSchema.ColumnCount)
+                    return GetScoreGetter(input, iinfo, activeOutput, out disposer);
+                else
+                    return GetFeatureContributionGetter(input, iinfo, activeOutput, out disposer);
+            }
         }
 
-        // TODO: .Transform() should be easy to implement -> just use the code inside create.
-        public IDataView Transform(IDataView input)
-        {
-            var roles = new List<KeyValuePair<RoleMappedSchema.ColumnRole, string>>();
-            roles.Add(new KeyValuePair<RoleMappedSchema.ColumnRole, string>(RoleMappedSchema.ColumnRole.Feature, _features));
-            var schema = new RoleMappedSchema(input.Schema, roles);
-
-            var boundMapper = _mapper.Bind(_host, schema);
-            return Create(_host, null, input, boundMapper, null);
-        }
-
-        public IRowToRowMapper GetRowToRowMapper(Schema inputSchema)
-        {
-            var roles = new List<KeyValuePair<RoleMappedSchema.ColumnRole, string>>();
-            roles.Add(new KeyValuePair<RoleMappedSchema.ColumnRole, string>(RoleMappedSchema.ColumnRole.Feature, _features));
-            var schema = new RoleMappedSchema(inputSchema, roles);
-
-            var boundMapper = _mapper.Bind(_host, schema);
-            return boundMapper as RowMapper;
-        }
-
-        private static IDataScorerTransform Create(IHostEnvironment env, Arguments args, IDataView data, ISchemaBoundMapper mapper, RoleMappedSchema trainSchema)
-        {
-            Contracts.CheckValue(env, nameof(env));
-            env.CheckValue(data, nameof(data));
-            env.CheckValue(mapper, nameof(mapper));
-
-            var contributionMapper = mapper as RowMapper;
-            env.CheckParam(mapper != null, nameof(mapper), "Unexpected mapper");
-
-            var scorer = ScoreUtils.GetScorerComponent(env, contributionMapper);
-            var scoredPipe = scorer.CreateComponent(env, data, contributionMapper, trainSchema);
-            return scoredPipe;
-        }
-
-        private static BindableMapper CreateBindableMapper(IHostEnvironment env, Arguments args, IPredictor predictor)
-        {
-            Contracts.CheckValue(env, nameof(env));
-            env.CheckValue(args, nameof(args));
-            env.CheckValue(predictor, nameof(predictor));
-            if (args.Top <= 0 || args.Top > MaxTopBottom)
-                throw env.Except($"Number of top contribution must be in range (0,{MaxTopBottom}]");
-            if (args.Bottom <= 0 || args.Bottom > MaxTopBottom)
-                throw env.Except($"Number of bottom contribution must be in range (0,{MaxTopBottom}]");
-
-            var pred = predictor as IFeatureContributionMapper;
-            env.CheckParam(pred != null, nameof(predictor), "Predictor doesn't support getting feature contributions");
-            return new BindableMapper(env, pred, args.Top, args.Bottom, args.Normalize, args.Stringify);
-        }
-
-        private static IDataScorerTransform Create(IHostEnvironment env, Arguments args, IDataView data, IPredictor predictor, string features = DefaultColumnNames.Features)
-        {
-            Contracts.CheckValue(env, nameof(env));
-            env.CheckValue(args, nameof(args));
-            env.CheckValue(predictor, nameof(predictor));
-            if (args.Top <= 0 || args.Top > MaxTopBottom)
-                throw env.Except($"Number of top contribution must be in range (0,{MaxTopBottom}]");
-            if (args.Bottom <= 0 || args.Bottom > MaxTopBottom)
-                throw env.Except($"Number of bottom contribution must be in range (0,{MaxTopBottom}]");
-
-            var roles = new List<KeyValuePair<RoleMappedSchema.ColumnRole, string>>();
-            roles.Add(new KeyValuePair<RoleMappedSchema.ColumnRole, string>(RoleMappedSchema.ColumnRole.Feature, features));
-            var schema = new RoleMappedSchema(data.Schema, roles);
-
-            var mapper = CreateBindableMapper(env, args, predictor);
-            var boundMapper = mapper.Bind(env, schema);
-            return Create(env, null, data, boundMapper, null);
-        }
-
-        private static ISchemaBindableMapper Create(IHostEnvironment env, Arguments args, IPredictor predictor)
-            => CreateBindableMapper(env, args, predictor);
-
-        private static ISchemaBindableMapper Create(IHostEnvironment env, ModelLoadContext ctx)
-            => new BindableMapper(env, ctx);
-
-        //private static IRowMapper Create(IHostEnvironment env, ModelLoadContext ctx, ISchema schema)
-        //{
-
-        //}
-
-        // TODO documentation... what is this? Can I put everything that is in here in the transformer, and eliminate this thing?
+        // TODO documentation
         internal sealed class BindableMapper : ISchemaBindableMapper, ICanSaveModel, IPredictor
         {
             private readonly int _topContributionsCount;
@@ -246,7 +281,7 @@ namespace Microsoft.ML.Runtime.Data
                     verReadableCur: 0x00010001,
                     verWeCanReadBack: 0x00010001,
                     loaderSignature: MapperLoaderSignature,
-                    loaderAssemblyName: typeof(FeatureContributionCalculationTransform).Assembly.FullName);
+                    loaderAssemblyName: typeof(FeatureContributionCalculatingTransformer).Assembly.FullName);
             }
 
             public PredictionKind PredictionKind => Predictor.PredictionKind;
@@ -270,6 +305,7 @@ namespace Microsoft.ML.Runtime.Data
                 GenericMapper = ScoreUtils.GetSchemaBindableMapper(_env, Predictor, null);
             }
 
+            // Factory method for SignatureLoadModel.
             public BindableMapper(IHostEnvironment env, ModelLoadContext ctx)
             {
                 Contracts.CheckValue(env, nameof(env));
@@ -292,12 +328,37 @@ namespace Microsoft.ML.Runtime.Data
                 Stringify = ctx.Reader.ReadBoolByte();
             }
 
+            // Factory method for SignatureDataScorer.
+            private static IDataScorerTransform Create(IHostEnvironment env, Arguments args, IDataView data, ISchemaBoundMapper mapper, RoleMappedSchema trainSchema)
+            {
+                Contracts.CheckValue(env, nameof(env));
+                env.CheckValue(data, nameof(data));
+                env.CheckValue(mapper, nameof(mapper));
+
+                var contributionMapper = mapper as BoundMapper;
+                env.CheckParam(mapper != null, nameof(mapper), "Unexpected mapper");
+
+                var scorer = ScoreUtils.GetScorerComponent(env, contributionMapper);
+                var scoredPipe = scorer.CreateComponent(env, data, contributionMapper, trainSchema);
+                return scoredPipe;
+            }
+
+            // Factory method for SignatureBindableMapper.
+            private static ISchemaBindableMapper Create(IHostEnvironment env, Arguments args, IPredictor predictor)
+            {
+                Contracts.CheckValue(env, nameof(env));
+                env.CheckValue(predictor, nameof(predictor));
+                var pred = predictor as IFeatureContributionMapper;
+                env.CheckParam(pred != null, nameof(predictor), "Predictor doesn't support getting feature contributions");
+                return new BindableMapper(env, pred, args.Top, args.Bottom, args.Normalize, args.Stringify);
+            }
+
             public ISchemaBoundMapper Bind(IHostEnvironment env, RoleMappedSchema schema)
             {
                 Contracts.CheckValue(env, nameof(env));
                 env.CheckValue(schema, nameof(schema));
                 CheckSchemaValid(env, schema, Predictor);
-                return new RowMapper(env, this, schema);
+                return new BoundMapper(env, this, schema);
             }
 
             public void Save(ModelSaveContext ctx)
@@ -428,105 +489,42 @@ namespace Microsoft.ML.Runtime.Data
             }
         }
 
-        private sealed class RowMapper : ISchemaBoundRowMapper
+        private sealed class BoundMapper : ISchemaBoundMapper
         {
             private readonly IHostEnvironment _env;
-            private readonly ISchemaBoundRowMapper _genericRowMapper;
-            private readonly BindableMapper _parent;
-            private readonly ISchema _outputSchema;
-            private readonly ISchema _outputGenericSchema;
-            private VBuffer<ReadOnlyMemory<char>> _slotNames;
 
             public RoleMappedSchema InputRoleMappedSchema { get; }
-
-            public Schema InputSchema => InputRoleMappedSchema.Schema;
-
             public Schema OutputSchema { get; }
+            public ISchemaBindableMapper Bindable { get; }
 
-            public ISchemaBindableMapper Bindable => _parent;
-
-            public RowMapper(IHostEnvironment env, BindableMapper parent, RoleMappedSchema schema)
+            public BoundMapper(IHostEnvironment env, BindableMapper parent, RoleMappedSchema schema)
             {
-                Contracts.AssertValue(env);
-                _env = env;
-                _env.AssertValue(schema);
-                _env.AssertValue(parent);
-                _env.AssertValue(schema.Feature);
-                _parent = parent;
+                Bindable = parent;
                 InputRoleMappedSchema = schema;
-                var genericMapper = parent.GenericMapper.Bind(_env, schema);
-                _genericRowMapper = genericMapper as ISchemaBoundRowMapper;
+                _env = env;
+
+                var inputSchema = schema.Schema;
+                var genericRowMapper = parent.GenericMapper.Bind(_env, schema) as ISchemaBoundRowMapper;
+                ISchema outputSchema;
 
                 if (parent.Stringify)
                 {
-                    _outputSchema = new SimpleSchema(_env,
+                    outputSchema = new SimpleSchema(_env,
                         new KeyValuePair<string, ColumnType>(DefaultColumnNames.FeatureContributions, TextType.Instance));
-                    if (InputSchema.HasSlotNames(InputRoleMappedSchema.Feature.Index, InputRoleMappedSchema.Feature.Type.VectorSize))
-                        InputSchema.GetMetadata(MetadataUtils.Kinds.SlotNames, InputRoleMappedSchema.Feature.Index,
-                            ref _slotNames);
-                    else
-                        _slotNames = VBufferUtils.CreateEmpty<ReadOnlyMemory<char>>(InputRoleMappedSchema.Feature.Type.VectorSize);
                 }
                 else
                 {
-                    _outputSchema = new FeatureContributionSchema(_env, DefaultColumnNames.FeatureContributions,
+                    outputSchema = new FeatureContributionSchema(_env, DefaultColumnNames.FeatureContributions,
                         new VectorType(NumberType.R4, schema.Feature.Type.AsVector),
-                        InputSchema, InputRoleMappedSchema.Feature.Index);
+                        inputSchema, InputRoleMappedSchema.Feature.Index);
                 }
-                _outputGenericSchema = _genericRowMapper.OutputSchema;
-                OutputSchema = new CompositeSchema(new ISchema[] { _outputGenericSchema, _outputSchema, }).AsSchema;
-            }
-
-            /// <summary>
-            /// Returns the input columns needed for the requested output columns.
-            /// </summary>
-            public Func<int, bool> GetDependencies(Func<int, bool> predicate)
-            {
-                for (int i = 0; i < OutputSchema.ColumnCount; i++)
-                {
-                    if (predicate(i))
-                        return col => col == InputRoleMappedSchema.Feature.Index;
-                }
-                return col => false;
-            }
-
-            public Row GetOutputRow(Row input, Func<int, bool> predicate, out Action disposer)
-            {
-                Contracts.AssertValue(input);
-                Contracts.AssertValue(predicate);
-                var totalColumnsCount = 1 + _outputGenericSchema.ColumnCount;
-                var getters = new Delegate[totalColumnsCount];
-
-                var genericRow = _genericRowMapper.GetRow(input, GetGenericPredicate(predicate), out disposer);
-                for (var i = 0; i < _outputGenericSchema.ColumnCount; i++)
-                {
-                    if (genericRow.IsColumnActive(i))
-                        getters[input.Schema.ColumnCount + i] = RowCursorUtils.GetGetterAsDelegate(genericRow, i);
-                }
-
-                if (predicate(totalColumnsCount - 1))
-                {
-                    getters[totalColumnsCount - 1] = _parent.Stringify
-                        ? _parent.GetTextContributionGetter(input, InputRoleMappedSchema.Feature.Index, _slotNames)
-                        : _parent.GetContributionGetter(input, InputRoleMappedSchema.Feature.Index);
-                }
-
-                return new SimpleRow(OutputSchema, input, getters);
-            }
-
-            public Func<int, bool> GetGenericPredicate(Func<int, bool> predicate)
-            {
-                return col => predicate(col);
+                ISchema outputGenericSchema = genericRowMapper.OutputSchema;
+                OutputSchema = new CompositeSchema(new ISchema[] { outputGenericSchema, outputSchema }).AsSchema;
             }
 
             public IEnumerable<KeyValuePair<RoleMappedSchema.ColumnRole, string>> GetInputColumnRoles()
             {
                 yield return RoleMappedSchema.ColumnRole.Feature.Bind(InputRoleMappedSchema.Feature.Name);
-            }
-
-            public Row GetRow(Row input, Func<int, bool> active, out Action disposer)
-            {
-                return GetOutputRow(input, active, out disposer);
             }
         }
 
@@ -600,6 +598,55 @@ namespace Microsoft.ML.Runtime.Data
                 else
                     throw MetadataUtils.ExceptGetMetadata();
             }
+        }
+    }
+
+    // TODO DOcumentation
+    public sealed class FeatureContributionCalculatingEstimator : TrivialEstimator<FeatureContributionCalculatingTransformer>
+    {
+        private readonly FeatureContributionCalculatingTransformer.Arguments _args;
+        private readonly string _features;
+        private readonly IPredictor _predictor;
+
+        // TODO Documentation
+        public FeatureContributionCalculatingEstimator(IHostEnvironment env, IPredictor predictor, string featuresColumn, FeatureContributionCalculatingTransformer.Arguments args)
+            : base(Contracts.CheckRef(env, nameof(env)).Register(nameof(FeatureContributionCalculatingTransformer)), new FeatureContributionCalculatingTransformer(env, predictor, featuresColumn, args))
+        {
+            // TODO argcheck
+            _args = args;
+            _features = featuresColumn;
+            _predictor = predictor;
+        }
+
+        public override SchemaShape GetOutputSchema(SchemaShape inputSchema)
+        {
+            Host.CheckValue(inputSchema, nameof(inputSchema));
+            var result = inputSchema.Columns.ToList();
+
+            if (!inputSchema.TryFindColumn(_features, out var col))
+                throw Host.ExceptSchemaMismatch(nameof(inputSchema), "input", _features);
+            var metadata = new List<SchemaShape.Column>();
+            if (col.Metadata.TryFindColumn(MetadataUtils.Kinds.SlotNames, out var slotMeta))
+                metadata.Add(slotMeta);
+            // TODO: check type of feature column.
+
+            // TODO: How do we deal with multiclassScoreColumn? should also contain slotnames
+            // Add Score column.
+            result.Add(new SchemaShape.Column(DefaultColumnNames.Score, SchemaShape.Column.VectorKind.Scalar, NumberType.R4, false, new SchemaShape(MetadataUtils.GetTrainerOutputMetadata())));
+
+            // Add FeatureContributions column.
+            if (_args.Stringify)
+            {
+                result.Add(new SchemaShape.Column(DefaultColumnNames.FeatureContributions, col.Kind,
+                    TextType.Instance, false, new SchemaShape(metadata.ToArray())));
+            }
+            else
+            {
+                result.Add(new SchemaShape.Column(DefaultColumnNames.FeatureContributions, col.Kind,
+                    col.ItemType, false, new SchemaShape(metadata.ToArray())));
+            }
+
+            return new SchemaShape(result.ToArray());
         }
     }
 }
